requirements.txt
apache-airflow==2.4.0